+++ BUILD.gn
@@ -80,13 +80,11 @@ rtc_shared_library("libwebrtc") {
     "include/rtc_sdp_parse_error.h",
     "include/rtc_ice_candidate.h",
     "include/rtc_media_stream.h",
-    "include/rtc_media_stream_list.h",
     "include/rtc_media_track.h",
     "include/rtc_mediaconstraints.h",
     "include/rtc_peerconnection.h",
     "include/rtc_peerconnection_factory.h",
     "include/rtc_rtp_capabilities.h",
-    "include/rtc_rtp_capabilities_lists.h",
     "include/rtc_rtp_parameters.h",
     "include/rtc_rtp_receiver.h",
     "include/rtc_rtp_sender.h",
@@ -126,8 +124,6 @@ rtc_shared_library("libwebrtc") {
     "src/rtc_ice_candidate_impl.h",
     "src/rtc_media_stream_impl.cc",
     "src/rtc_media_stream_impl.h",
-    "src/rtc_media_stream_list_impl.cc",
-    "src/rtc_media_stream_list_impl.h",
     "src/rtc_mediaconstraints_impl.cc",
     "src/rtc_mediaconstraints_impl.h",
     "src/rtc_peerconnection_factory_impl.cc",
@@ -136,8 +132,6 @@ rtc_shared_library("libwebrtc") {
     "src/rtc_peerconnection_impl.h",
     "src/rtc_rtp_capabilities_impl.cc",
     "src/rtc_rtp_capabilities_impl.h",
-    "src/rtc_rtp_capabilities_lists_impl.cc",
-    "src/rtc_rtp_capabilities_lists_impl.h",
     "src/rtc_rtp_parameters_impl.cc",
     "src/rtc_rtp_parameters_impl.h",
     "src/rtc_rtp_receiver_impl.cc",
@@ -181,7 +175,6 @@ rtc_shared_library("libwebrtc") {
       "src/interop/rtc_data_channel_interop.cc",
       "src/interop/rtc_ice_candidate_interop.cc",
       "src/interop/rtc_media_stream_interop.cc",
-      "src/interop/rtc_media_stream_list_interop.cc",
       "src/interop/rtc_rtp_capabilities_interop.cc",
       "src/interop/rtc_rtp_transceiver_interop.cc",
       "src/interop/rtc_peerconnection_interop.cc",
